adopcion-backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
mysqlclient==2.2.6
python-decouple==3.8
Pillow==11.0.0
